feat: Реализация функции для поиска символа в строке
